+++ Exception/RateLimitExceededException.php
@@ -18,9 +18,9 @@ use Symfony\Component\HttpKernel\Exception\HttpException;
  */
 class RateLimitExceededException extends HttpException
 {
-    public function __construct(int $statusCode, string $message, string $ip)
+    public function __construct(int $statusCode, string $message, string $ip, string $username = null)
     {
-        $message = sprintf($message, $ip);
+        $message = sprintf($message, $username ?: $ip);
 
         parent::__construct($statusCode, $message);
     }