+++ Tests/Service/RateLimitHandlerTest.php
@@ -18,6 +18,11 @@ use PHPUnit\Framework\TestCase;
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
+use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
+use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationChecker;
 
 class RateLimitHandlerTest extends TestCase
 {
@@ -32,15 +37,23 @@ class RateLimitHandlerTest extends TestCase
         $cacheItemPool->expects($this->never())
             ->method('getItem');
 
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $throttleConfig = [
-            'limit' => 60,
-            'period' => 60,
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [],
         ];
 
         $request = Request::create('/api/me');
         $request->attributes->set('_api_resource_class', DisableRateLimit::class);
 
-        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $throttleConfig);
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
         $rateLimitHandler->handle($request);
 
         $this->assertFalse($rateLimitHandler->isEnabled());
@@ -59,15 +72,23 @@ class RateLimitHandlerTest extends TestCase
             ->method('getItem')
             ->will($this->returnValue($cacheItem));
 
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $throttleConfig = [
-            'limit' => 60,
-            'period' => 60,
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [],
         ];
 
         $request = Request::create('/api/me');
         $request->attributes->set('_api_resource_class', EnableRateLimit::class);
 
-        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $throttleConfig);
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
         $rateLimitHandler->handle($request);
 
         $this->assertTrue($rateLimitHandler->isEnabled());
@@ -80,8 +101,6 @@ class RateLimitHandlerTest extends TestCase
 
     public function testRateLimitIsExceeded()
     {
-        $resetTime = gmdate('U');
-
         $cacheItem = $this->createMock(CacheItemInterface::class);
 
         $cacheItem->expects($this->once())
@@ -93,7 +112,7 @@ class RateLimitHandlerTest extends TestCase
             ->will($this->returnValue([
                 'limit' => 60,
                 'remaining' => 0,
-                'reset' => $resetTime,
+                'reset' => gmdate('U'),
             ]));
 
         $cacheItemPool = $this->createMock(CacheItemPoolInterface::class);
@@ -101,15 +120,23 @@ class RateLimitHandlerTest extends TestCase
             ->method('getItem')
             ->will($this->returnValue($cacheItem));
 
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $throttleConfig = [
-            'limit' => 60,
-            'period' => 60,
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [],
         ];
 
         $request = Request::create('/api/me');
         $request->attributes->set('_api_resource_class', EnableRateLimit::class);
 
-        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $throttleConfig);
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
         $rateLimitHandler->handle($request);
 
         $this->assertTrue($rateLimitHandler->isEnabled());
@@ -121,8 +148,6 @@ class RateLimitHandlerTest extends TestCase
 
     public function testRateLimitIsNotExceeded()
     {
-        $resetTime = gmdate('U');
-
         $cacheItem = $this->createMock(CacheItemInterface::class);
 
         $cacheItem->expects($this->once())
@@ -134,7 +159,7 @@ class RateLimitHandlerTest extends TestCase
             ->will($this->returnValue([
                 'limit' => 60,
                 'remaining' => 50,
-                'reset' => $resetTime,
+                'reset' => gmdate('U'),
             ]));
 
         $cacheItemPool = $this->createMock(CacheItemPoolInterface::class);
@@ -142,15 +167,23 @@ class RateLimitHandlerTest extends TestCase
             ->method('getItem')
             ->will($this->returnValue($cacheItem));
 
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $throttleConfig = [
-            'limit' => 60,
-            'period' => 60,
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [],
         ];
 
         $request = Request::create('/api/me');
         $request->attributes->set('_api_resource_class', EnableRateLimit::class);
 
-        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $throttleConfig);
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
         $rateLimitHandler->handle($request);
 
         $this->assertTrue($rateLimitHandler->isEnabled());
@@ -158,6 +191,135 @@ class RateLimitHandlerTest extends TestCase
         $rateLimitInfo = $rateLimitHandler->getRateLimitInfo();
         $this->assertSame(60, $rateLimitInfo['limit']);
         $this->assertSame(49, $rateLimitInfo['remaining']);
+
+        $this->assertFalse($rateLimitHandler->isRateLimitExceeded());
+    }
+
+    public function testRateLimitByRole()
+    {
+        $cacheItem = $this->createMock(CacheItemInterface::class);
+
+        $cacheItem->expects($this->once())
+            ->method('isHit')
+            ->will($this->returnValue(false));
+
+        $cacheItem->expects($this->once())
+            ->method('get')
+            ->will($this->returnValue([]));
+
+        $cacheItemPool = $this->createMock(CacheItemPoolInterface::class);
+        $cacheItemPool->expects($this->once())
+            ->method('getItem')
+            ->will($this->returnValue($cacheItem));
+
+        $token = $this->getMockBuilder(UsernamePasswordToken::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $token->expects($this->once())
+            ->method('isAuthenticated')
+            ->will($this->returnValue(true));
+
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $tokenStorage->expects($this->any())
+            ->method('getToken')
+            ->will($this->returnValue($token));
+
+        $authenticationManager = $this->createMock(AuthenticationManagerInterface::class);
+        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager->expects($this->once())
+            ->method('decide')
+            ->will($this->returnValue(true));
+
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->setConstructorArgs([$tokenStorage, $authenticationManager, $accessDecisionManager])
+            ->getMock();
+
+        $throttleConfig = [
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [
+                'ROLE_USER' => [
+                    'limit' => 10,
+                    'period' => 10,
+                ],
+            ],
+        ];
+
+        $request = Request::create('/api/me');
+        $request->attributes->set('_api_resource_class', EnableRateLimit::class);
+
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
+        $rateLimitHandler->handle($request);
+
+        $this->assertTrue($rateLimitHandler->isEnabled());
+
+        $rateLimitInfo = $rateLimitHandler->getRateLimitInfo();
+        $this->assertSame(10, $rateLimitInfo['limit']);
+        $this->assertSame(9, $rateLimitInfo['remaining']);
+
+        $this->assertFalse($rateLimitHandler->isRateLimitExceeded());
+    }
+
+    public function testAuthenticationCredentialNotFound()
+    {
+        $cacheItem = $this->createMock(CacheItemInterface::class);
+
+        $cacheItem->expects($this->once())
+            ->method('isHit')
+            ->will($this->returnValue(false));
+
+        $cacheItem->expects($this->once())
+            ->method('get')
+            ->will($this->returnValue([]));
+
+        $cacheItemPool = $this->createMock(CacheItemPoolInterface::class);
+        $cacheItemPool->expects($this->once())
+            ->method('getItem')
+            ->will($this->returnValue($cacheItem));
+
+        $token = $this->getMockBuilder(UsernamePasswordToken::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $tokenStorage = $this->createMock(TokenStorage::class);
+        $tokenStorage->expects($this->any())
+            ->method('getToken')
+            ->will($this->returnValue(null));
+
+        $authenticationManager = $this->createMock(AuthenticationManagerInterface::class);
+        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+
+        $authorizationChecker = $this->getMockBuilder(AuthorizationChecker::class)
+            ->setConstructorArgs([$tokenStorage, $authenticationManager, $accessDecisionManager])
+            ->getMock();
+
+        $throttleConfig = [
+            'default' => [
+                'limit' => 60,
+                'period' => 60,
+            ],
+            'roles' => [
+                'ROLE_USER' => [
+                    'limit' => 10,
+                    'period' => 10,
+                ],
+            ],
+        ];
+
+        $request = Request::create('/api/me');
+        $request->attributes->set('_api_resource_class', EnableRateLimit::class);
+
+        $rateLimitHandler = new RateLimitHandler($cacheItemPool, $tokenStorage, $authorizationChecker, $throttleConfig);
+        $rateLimitHandler->handle($request);
+
+        $this->assertTrue($rateLimitHandler->isEnabled());
+
+        $rateLimitInfo = $rateLimitHandler->getRateLimitInfo();
+        $this->assertSame(60, $rateLimitInfo['limit']);
+        $this->assertSame(59, $rateLimitInfo['remaining']);
 
         $this->assertFalse($rateLimitHandler->isRateLimitExceeded());
     }