+++ CHANGELOG.md
@@ -4,4 +4,7 @@ CHANGELOG
 master
 ------
 
-* deprecated the `indragunawan_api_rate_limit.storage` configuration key.
+* Feature. Using PSR/Cache instead of DoctrineCache (#2).
+* Feature. Role based rate limit. (#3)
+* Deprecated the `indragunawan_api_rate_limit.storage` configuration key.
+* Deprecated the `indragunawan_api_rate_limit.throttle.period` and `indragunawan_api_rate_limit.throttle.limit` configuration key.